move generate-pair-name into /domain-service
