+++ src/config/participant-router.ts
@@ -8,7 +8,7 @@ import { TeamRepository } from "../infra/team/team-repository"
 import { CheckEmailAlreadyExists } from "../domain/participant/check-email-already-exists"
 import { ParticipantFactory } from "../domain/participant/participant-factory"
 import { PairFactory } from "../domain/pair/pair-factory"
-import { GeneratePairName } from "../domain/pair/generate-pair-name"
+import { GeneratePairName } from "../domain/pair/domain-service/generate-pair-name"
 import { JoinPair } from "../domain/pair/join-pair"
 import { GetVacantPairList } from "../domain/pair/get-vacant-pair-list"
 import { GetParentTeam } from "../domain/pair/get-parent-team"
