+++ src/domain/pair/join-pair.ts
@@ -6,7 +6,7 @@ import { IPairRepository } from "./interface/pair-repository"
 import { PairFactory } from "./pair-factory"
 import { GetVacantPairList } from "./get-vacant-pair-list"
 import { GetParentTeam } from "./get-parent-team"
-import { GeneratePairName } from "./generate-pair-name"
+import { GeneratePairName } from "./domain-service/generate-pair-name"
 
 export class JoinPair extends DomainService<"join-pair"> {
   constructor(
